% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{all_motif_datasets}
\alias{all_motif_datasets}
\title{Get a data frame of all the motif datasets bundled with prego}
\usage{
all_motif_datasets()
}
\value{
a data frame which concatenates motifs from "HOMER", "JASPAR" and "JOLMA".
Motif names are prefixed with the dataset name, e.g. "JASPAR.GATA4".
}
\description{
Get a data frame of all the motif datasets bundled with prego
}
\examples{
all_motif_datasets()

}
