% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{as.data.frame,MotifDB-method}
\alias{as.data.frame,MotifDB-method}
\title{Convert a MotifDB object to a data frame}
\usage{
\S4method{as.data.frame}{MotifDB}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A MotifDB object}

\item{row.names}{NULL or a character vector giving the row names for the data frame}

\item{optional}{logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional}

\item{...}{additional arguments to be passed to or from methods}
}
\value{
A data frame containing the motif probabilities
}
\description{
Convert a MotifDB object to a data frame
}
