% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{as.matrix,MotifDB-method}
\alias{as.matrix,MotifDB-method}
\title{Convert a MotifDB object to a matrix}
\usage{
\S4method{as.matrix}{MotifDB}(x, ...)
}
\arguments{
\item{x}{MotifDB object}
}
\value{
A matrix containing the motif probabilities, rownames are motif_position, colnames are nucleotides
}
\description{
Convert a MotifDB object to a matrix
}
