% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm.R
\name{calc_seq_pwm}
\alias{calc_seq_pwm}
\title{Calculate Position Weight Matrix (PWM) Scores for DNA Sequences}
\usage{
calc_seq_pwm(sequences, mdb, bidirect = TRUE)
}
\arguments{
\item{sequences}{Character vector of DNA sequences.}

\item{mdb}{MotifDB object containing PWMs.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}
}
\value{
A numeric matrix with sequences as rows and motifs as columns, containing PWM scores.
Row names are preserved from input sequences if they exist.
Column names are preserved from the PWM matrix if they exist.
}
\description{
Calculate Position Weight Matrix (PWM) Scores for DNA Sequences
}
\examples{
sequences <- c("ACGTACGT", "TGCATGCA")
scores <- calc_seq_pwm(sequences, MOTIF_DB)

}
