% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{calc_sequences_dinucs}
\alias{calc_sequences_dinucs}
\title{Calculate Dinucleotide Counts for Sequences}
\usage{
calc_sequences_dinucs(sequences)
}
\arguments{
\item{sequences}{A character vector of DNA sequences. Each element should be
a string representing a DNA sequence composed of A, T, C, and G.}
}
\value{
A numeric matrix where:
\itemize{
\item Each row corresponds to a sequence in the input vector.
\item Each column represents a specific dinucleotide (AA, AC, AG, AT, CA, CC, etc.).
\item The values in the matrix are the counts of each dinucleotide in each sequence.
\item Column names are set to the corresponding dinucleotides.
}
}
\description{
This function calculates the total count of each dinucleotide for each sequence
in a vector of DNA sequences.
}
\examples{

sequences <- c("ATCG", "GCTA", "AATT")
result <- calc_sequences_dinucs(sequences)
print(result)

}
