% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial.R
\name{calc_sequences_trinuc_dist}
\alias{calc_sequences_trinuc_dist}
\title{Calculate Trinucleotide Distribution in Sequences}
\usage{
calc_sequences_trinuc_dist(sequences, size = NULL)
}
\arguments{
\item{sequences}{a character vector containing the sequences to analyze.
Each element of the vector should be a single sequence.}

\item{size}{an integer specifying the size to consider for the analysis.
If NULL (default), the maximum length of the sequences in the \code{sequences}
vector is used.}
}
\value{
a data frame with columns 'pos' and 64 columns representing each possible
trinucleotide. Each row represents a position in the sequences (from 1 to \code{size}),
and contains the fraction of each trinucleotide at that position across
all sequences.
}
\description{
Calculate Trinucleotide Distribution in Sequences
}
\examples{

# Generate some random sequences for testing
set.seed(60427)
sequences <- sapply(1:100, function(x) {
    paste0(sample(c("A", "C", "G", "T"), 1000, replace = TRUE), collapse = "")
})
sequences <- as.character(sequences)

# Calculate the trinucleotide distribution
result <- calc_sequences_trinuc_dist(sequences)

head(result)

}
