% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{cluster_mat_example}
\alias{cluster_mat_example}
\alias{cluster_sequences_example}
\alias{clusters_example}
\alias{response_mat_example}
\alias{sequences_example}
\title{Dataset examples}
\format{
sequences and response variables in various formats for function examples

An object of class \code{character} of length 2359.

An object of class \code{character} of length 2359.

An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 5 columns.

An object of class \code{character} of length 1000.
}
\usage{
cluster_mat_example

cluster_sequences_example

clusters_example

response_mat_example

sequences_example
}
\description{
Dataset examples
}
\keyword{datasets}
