% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{compute_local_pwm}
\alias{compute_local_pwm}
\title{Compute local PWMs for a set of sequences given a PSSM matrix}
\usage{
compute_local_pwm(
  sequences,
  pssm,
  spat = NULL,
  spat_min = 0,
  spat_max = NULL,
  bidirect = TRUE,
  prior = 0
)
}
\arguments{
\item{sequences}{a vector of sequences}

\item{pssm}{a PSSM matrix or data frame. The columns of the matrix or data frame should be named with the nucleotides ('A', 'C', 'G' and 'T').}

\item{spat}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}

\item{spat_min}{the minimum position to use from the sequences. The default is 1.}

\item{spat_max}{the maximum position to use from the sequences. The default is the length of the sequences.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{prior}{a prior probability for each nucleotide.}
}
\value{
a matrix with \code{length(sequences)} rows and \code{ncol(pssm)} columns with the local PWM for each sequence in each position.
}
\description{
compute the local PWM for each position in every sequence. The edges of each sequences would become NA.
}
\examples{
\dontrun{
res <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1])

pwm <- compute_local_pwm(cluster_sequences_example, res$pssm, res$spat)
head(pwm)
}

}
