% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{consensus_from_pssm}
\alias{consensus_from_pssm}
\title{Convert PSSM to consensus sequence}
\usage{
consensus_from_pssm(pssm, single_thresh = 0.6, double_thresh = 0.85)
}
\arguments{
\item{pssm}{A PSSM matrix}

\item{single_thresh, double_thresh}{thresholds for the consensus sequence calculation
(single and double nucleotides)}
}
\value{
A consensus sequence for the PSSM. If no consensus sequence can be found, the function returns NA.
}
\description{
Convert PSSM to consensus sequence
}
\examples{
\dontrun{
res <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1])
consensus_from_pssm(res$pssm)
}

}
