% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{create_motif_db}
\alias{create_motif_db}
\title{Create a MotifDB object from a tidy data frame}
\usage{
create_motif_db(motif_db, prior = 0.01)
}
\arguments{
\item{motif_db}{A tidy data frame containing motif information}

\item{prior}{Pseudocount prior to add to probabilities (default: 0.01)}
}
\value{
A MotifDB object
}
\description{
Create a MotifDB object from a tidy data frame
}
