% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_multi_regression}
\alias{export_multi_regression}
\title{Export a multiple motif regression model}
\usage{
export_multi_regression(reg, fn = NULL)
}
\arguments{
\item{reg}{a multiple motif regression model, as returned by \code{regress_pwm} with \code{motif_num > 1}}

\item{fn}{a file name to save the model to. If NULL - the model is returned as a list}
}
\value{
None
}
\description{
Export a multiple motif regression model
}
\examples{
\dontrun{
res_multi <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1],
    final_metric = "ks", spat_bin_size = 40,
    spat_num_bins = 7,
    motif_num = 2
)
export_fn <- tempfile()
export_multi_regression(res_multi, export_fn)

light_res <- export_multi_regression(res_multi)

# loading can be done by:
r <- load_multi_regression(export_fn)
}

}
