% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_regression_model}
\alias{export_regression_model}
\title{Export a motif regression model}
\usage{
export_regression_model(model, fn = NULL)
}
\arguments{
\item{model}{a motif regression model, as returned by \code{regress_pwm} with \code{motif_num = 1}}

\item{fn}{a file name to save the model to. If NULL - the model is returned as a list}
}
\value{
None
}
\description{
Export a motif regression model
}
\examples{
\dontrun{
res <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1],
    final_metric = "ks", spat_bin_size = 40,
    spat_num_bins = 7
)
export_fn <- tempfile()
export_regression_model(export_fn)
r <- load_regression(export_fn)
}

}
