% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-dbs.R
\name{extract_pwm}
\alias{extract_pwm}
\title{Extract pwm of sequences from a motif database}
\usage{
extract_pwm(
  sequences,
  motifs = NULL,
  dataset = all_motif_datasets(),
  spat = NULL,
  spat_min = 0,
  spat_max = NULL,
  bidirect = TRUE,
  parallel = getOption("prego.parallel", TRUE)
)
}
\arguments{
\item{sequences}{a vector of sequences}

\item{motifs}{names of specific motifs to extract from the dataset}

\item{dataset}{a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example \code{HOMER_motifs} or \code{JASPAR_motifs}, or \code{all_motif_datasets()}.}

\item{spat}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}

\item{spat_min}{start of the spatial model from the beginning of the sequence (in bp)}

\item{spat_max}{end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model
would end at the end of the sequence.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{parallel}{logical, whether to use parallel processing}
}
\value{
a matrix size of # of sequences x # of motifs with the pwm of each sequence for each motif
}
\description{
Extracts the pwm of a motif from a motif database. Note that for a large number of motifs, this function can be slow and consume
a lot of memory.
}
\examples{
pwms <- extract_pwm(cluster_sequences_example, motifs = c("JASPAR.CDX1", "HOMER.Hnf1", "HOMER.GATA3_2"))
head(pwms)

# all motifs
all_pwms <- extract_pwm(cluster_sequences_example)
dim(all_pwms)
all_pwms[1:5, 1:5]

# for a specific dataset
pwms_jaspar <- extract_pwm(cluster_sequences_example, dataset = JASPAR_motifs)
head(pwms_jaspar)

}
