% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer-regression.R
\name{generate_kmers}
\alias{generate_kmers}
\title{Generate kmers}
\usage{
generate_kmers(kmer_length, max_gap = 0, min_gap = 0)
}
\arguments{
\item{kmer_length}{The number of non-gap characters in motifs that will be screened.}

\item{max_gap}{The maximum length of a gap to be considered in the pattern. Default: 0}

\item{min_gap}{The minimum length of a gap to be considered in the pattern. Default: 0}
}
\value{
A vector of all possible kmers considering the gap length.
}
\description{
This function generates all possible kmers considering the gap length. Gaps are represented by 'N'.
}
\examples{

# Generate kmers of length 2 without any gaps
generate_kmers(2)

# Generate kmers of length 3 with a single gap (1 'N') at any position
generate_kmers(3, min_gap = 1, max_gap = 1)

# Generate kmers of length 3 with a gap of 1 to 2 'N's at any position
generate_kmers(3, min_gap = 1, max_gap = 2)

# Generate kmers of length 3 with a gap of 2 'N's at any position
generate_kmers(3, min_gap = 2, max_gap = 2)

}
