% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-dbs.R
\name{get_motif_pssm}
\alias{get_motif_pssm}
\title{Extract pssm of sequences from a motif database}
\usage{
get_motif_pssm(motif, dataset = all_motif_datasets())
}
\arguments{
\item{motif}{name of the motif to extract from the dataset}

\item{dataset}{a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example \code{HOMER_motifs} or \code{JASPAR_motifs}, or \code{all_motif_datasets()}.}
}
\value{
a data frame with the pssm of the motif
}
\description{
Extract pssm of sequences from a motif database
}
\examples{
get_motif_pssm("JASPAR.HNF1A")

}
