% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misha.R
\name{gextract_pwm}
\alias{gextract_pwm}
\title{Extract pwm of intervals from a motif database}
\usage{
gextract_pwm(
  intervals,
  motifs = NULL,
  dataset = all_motif_datasets(),
  spat = NULL,
  spat_min = 1,
  spat_max = NULL,
  bidirect = TRUE,
  prior = 0.01,
  func = "logSumExp",
  parallel = getOption("prego.parallel", TRUE)
)
}
\arguments{
\item{intervals}{misha intervals set}

\item{motifs}{names of specific motifs to extract from the dataset}

\item{dataset}{a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example \code{HOMER_motifs} or \code{JASPAR_motifs}, or \code{all_motif_datasets()}.}

\item{spat}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}

\item{spat_min}{the minimum position to use from the sequences. The default is 1.}

\item{spat_max}{the maximum position to use from the sequences. The default is the length of the sequences.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{prior}{a prior probability for each nucleotide.}

\item{func}{the function to use to combine the PWMs for each sequence. Either 'logSumExp' or 'max'. The default is 'logSumExp'.}

\item{parallel}{logical, whether to use parallel processing}
}
\value{
The intervals set with additional columns per motif, containing the pwm of each interval for each motif
}
\description{
Extract pwm of intervals from a motif database
}
\examples{
\dontrun{
library(misha)
gdb.init_examples()
pwms <- gextract_pwm("annotations")
pwms[, 1:20]
}

}
