% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{gintervals.center_by_pssm}
\alias{gintervals.center_by_pssm}
\title{Center intervals by PSSM}
\usage{
gintervals.center_by_pssm(
  intervals,
  pssm,
  spat = NULL,
  spat_min = 0,
  spat_max = NULL,
  bidirect = TRUE,
  prior = 0.01
)
}
\arguments{
\item{intervals}{The intervals to extract}

\item{pssm}{a PSSM matrix or data frame. The columns of the matrix or data frame should be named with the nucleotides ('A', 'C', 'G' and 'T').}

\item{spat}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}

\item{spat_min}{the minimum position to use from the sequences. The default is 1.}

\item{spat_max}{the maximum position to use from the sequences. The default is the length of the sequences.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{prior}{a prior probability for each nucleotide.}
}
\value{
A data frame containing the centered intervals. The intervals will have the same columns as the input
intervals, but the start and end positions will be adjusted to center the intervals based on the maximum score
position in the PSSM.
}
\description{
This function takes a set of intervals and a position-specific scoring matrix (PSSM) and centers the intervals
based on the maximum score position in the PSSM. The intervals are shifted so that the maximum score position
becomes the center of each interval.
}
