% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-dbs.R
\docType{data}
\name{HOCOMOCO_motifs}
\alias{HOCOMOCO_motifs}
\title{PSSMs from the HOCOMOCO motif database}
\format{
A data frame containing the PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name.
}
\source{
\url{https://hocomoco11.autosome.ru/downloads/}
}
\usage{
HOCOMOCO_motifs
}
\description{
PSSMs from the HOCOMOCO motif database
}
\references{
Ivan V. Kulakovskiy; Ilya E. Vorontsov; Ivan S. Yevshin; Ruslan N. Sharipov; Alla D. Fedorova; Eugene I. Rumynskiy; Yulia A. Medvedeva; Arturo Magana-Mora; Vladimir B. Bajic; Dmitry A. Papatsenko; Fedor A. Kolpakov; Vsevolod J. Makeev: HOCOMOCO: towards a complete collection of transcription factor binding models for human and mouse via large-scale ChIP-Seq analysis. Nucl. Acids Res., Database issue, gkx1106 (11 November 2017). \url{https://doi.org/10.1093/nar/gkx1106}
}
\keyword{datasets}
