% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-dbs.R
\docType{data}
\name{HOMER_motifs}
\alias{HOMER_motifs}
\title{PSSMs from the HOMER motif database}
\format{
A data frame containing the PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name.
}
\source{
\url{http://homer.ucsd.edu/homer/motif/}
}
\usage{
HOMER_motifs
}
\description{
PSSMs from the HOMER motif database
}
\references{
Heinz S, Benner C, Spann N, Bertolino E et al. Simple Combinations of Lineage-Determining Transcription Factors Prime cis-Regulatory Elements Required for Macrophage and B Cell Identities. Mol Cell 2010 May 28;38(4):576-589. PMID: 20513432
}
\keyword{datasets}
