% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{initialize,MotifDB-method}
\alias{initialize,MotifDB-method}
\title{Initialize method for MotifDB objects}
\usage{
\S4method{initialize}{MotifDB}(
  .Object,
  mat = matrix(0, 4, 1),
  rc_mat = matrix(0, 4, 1),
  motif_lengths = setNames(as.integer(1), colnames(mat)[1]),
  prior = 0.01
)
}
\arguments{
\item{.Object}{MotifDB object}

\item{mat}{Position weight matrix}

\item{rc_mat}{Reverse complement matrix}

\item{motif_lengths}{Named vector of motif lengths}

\item{prior}{Pseudocount prior value}
}
\description{
Initialize method for MotifDB objects
}
