% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misha.R
\name{intervals_to_seq}
\alias{intervals_to_seq}
\title{Convert intervals to sequences}
\usage{
intervals_to_seq(intervals)
}
\arguments{
\item{intervals}{The intervals set as a data frame with 'chrom', 'start', and 'end' columns.}
}
\value{
A character vector of sequences.
}
\description{
This function takes a set of intervals and converts them into sequences.
It requires the 'misha' package to be installed. If the package is not
installed, it will display an error message with instructions on how to
install it.
}
\examples{
\dontrun{
library(misha)
gdb.init_examples()
intervals_to_seq(gintervals.load("annotations"))
}
}
