% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-dbs.R
\docType{data}
\name{JASPAR_motifs}
\alias{JASPAR_motifs}
\title{PSSMs from the JASPAR motif database}
\format{
A data frame containing the PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name.
}
\source{
\url{https://jaspar.elixir.no/downloads/}
}
\usage{
JASPAR_motifs
}
\description{
PSSMs from the JASPAR motif database
}
\references{
Castro-Mondragon JA, Riudavets-Puig R, Rauluseviciute I, Berhanu Lemma R, Turchi L, Blanc-Mathieu R, Lucas J, Boddie P, Khan A, Manosalva Pérez N, Fornes O, Leung TY, Aguirre A, Hammal F, Schmelter D, Baranasic D, Ballester B, Sandelin A, Lenhard B, Vandepoele K, Wasserman WW, Parcy F, and Mathelier A JASPAR 2022: the 9th release of the open-access database of transcription factor binding profiles Nucleic Acids Res. 2022 Jan 7;50(D1):D165-D173.; doi: 10.1093/nar/gkab1113
}
\keyword{datasets}
