% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer-regression.R
\name{kmers_to_pssm}
\alias{kmers_to_pssm}
\title{Transform k-mers to PSSM (Position-Specific Scoring Matrix)}
\usage{
kmers_to_pssm(kmers, prior = 0.01)
}
\arguments{
\item{kmers}{A character vector of k-mers.}

\item{prior}{A numeric value indicating the prior probability for each nucleotide. Default is 0.01.}
}
\value{
A data frame with columns for the k-mer, position, and nucleotide frequencies, 'kmer', 'pos', 'A', 'C', 'G', 'T'.
}
\description{
This function transforms a vector of k-mers into a position-specific scoring matrix (PSSM).
A PSSM represents the frequency of each nucleotide at each position in the k-mers.
If a nucleotide is 'N', it is treated as equal probabilities for 'A', 'C', 'G', and 'T'.
The result is returned as a data frame with columns for the k-mer, position, and nucleotide frequencies.
}
\examples{
kmers_to_pssm(c("ACGTN", "TGCAN"), prior = 0.01)

}
