% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{length,MotifDB-method}
\alias{length,MotifDB-method}
\title{Get the length of a MotifDB object}
\usage{
\S4method{length}{MotifDB}(x)
}
\arguments{
\item{x}{MotifDB object}
}
\value{
The number of motifs in the object
}
\description{
Get the length of a MotifDB object
}
\examples{
length(MOTIF_DB)
length(MOTIF_DB[c("HOMER.GATA3_2", "JASPAR.CDX1")])
}
