% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{load_multi_regression}
\alias{load_multi_regression}
\title{Load a multiple motif regression model from a file}
\usage{
load_multi_regression(
  fn,
  response = NULL,
  sequences = NULL,
  motif_dataset = all_motif_datasets(),
  parallel = getOption("prego.parallel", FALSE),
  alternative = "two.sided"
)
}
\arguments{
\item{fn}{file name or a list with the model}

\item{response}{A matrix (or vector) of response variables. The number of rows must equal the number of sequences. A single binary vector (0/1) is required when \code{score_metric = "ks"}.}

\item{sequences}{A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through \code{toupper}). Sequences must be long enough to cover \code{spat_num_bins} * \code{spat_bin_size} bp, and should be centered around the motif/signal.}

\item{motif_dataset}{A data frame with PSSMs (columns \code{A}, \code{C}, \code{G}, \code{T}, \code{motif}), e.g. \code{HOMER_motifs}, \code{JASPAR_motifs}, or \code{all_motif_datasets()} (default).}

\item{parallel}{Whether to parallelize. Use \code{set_parallel} to set the number of cores.}

\item{alternative}{Alternative hypothesis for \code{ks.test}. One of \code{"two.sided"}, \code{"less"}, \code{"greater"}.}
}
\value{
a list with the following elements:

\describe{
\item{models: }{a list of models.}
\item{model: }{the combined model.}
\item{spat_min: }{the minimum spatial position.}
\item{spat_max: }{the maximum spatial position.}
\item{bidirect: }{whether the model is bidirectional.}
\item{spat_bin_size: }{the spatial bin size.}
\item{seq_length: }{the sequence length.}
\item{motif_num: }{the number of motifs.}
\item{predict: }{a function to predict the response.}
\item{predict_multi: }{a function to predict the response for each motif.}
}
}
\description{
Load a multiple motif regression model from a file
}
\examples{
\dontrun{
res_multi <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1],
    final_metric = "ks", spat_bin_size = 40,
    spat_num_bins = 7,
    motif_num = 2
)
tmp <- tempfile()
res_multi$export(tmp)
r <- load_multi_regression(tmp)
}

}
