% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{load_regression_model}
\alias{load_regression_model}
\title{Load a motif regression model from a file}
\usage{
load_regression_model(fn)
}
\arguments{
\item{fn}{file name or a list with the model}
}
\value{
a list with the following elements:

\describe{
\item{pssm:}{a data frame with the PSSM}
\item{spat:}{a data frame with the spatial profile}
\item{spat_min:}{a numeric value with the minimum value of the spatial profile}
\item{spat_max:}{a numeric value with the maximum value of the spatial profile}
\item{bidirect:}{a boolean value indicating whether the model is bidirectional}
\item{seq_length:}{a numeric value with the length of the sequences}
}
}
\description{
Load a motif regression model from a file
}
\examples{
\dontrun{
res <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1],
    final_metric = "ks", spat_bin_size = 40,
    spat_num_bins = 7
)
export_fn <- tempfile()
export_regression_model(export_fn)
r <- load_regression(export_fn)
}

}
