% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{mask_sequences_by_pwm}
\alias{mask_sequences_by_pwm}
\title{Mask sequences by thresholding the PWM}
\usage{
mask_sequences_by_pwm(
  sequences,
  pssm,
  mask_thresh,
  pos_bits_thresh = 0.2,
  spat = NULL,
  spat_min = 0,
  spat_max = NULL,
  bidirect = TRUE,
  prior = 0
)
}
\arguments{
\item{sequences}{a vector of sequences}

\item{pssm}{a PSSM matrix or data frame. The columns of the matrix or data frame should be named with the nucleotides ('A', 'C', 'G' and 'T').}

\item{mask_thresh}{Threshold for masking. Sequences with a PWM above this threshold will be masked by 'N'.}

\item{pos_bits_thresh}{Mask only positions with amount of information contributed (Shannon entropy, measured in bits) above this threshold.
The scale is the same as the y axis in the pssm logo plots.}

\item{spat}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}

\item{spat_min}{start of the spatial model from the beginning of the sequence (in bp)}

\item{spat_max}{end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model
would end at the end of the sequence.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{prior}{a prior probability for each nucleotide.}
}
\value{
A vector with the masked sequences.
}
\description{
Mask sequences by thresholding the PWM. Sequences with a PWM above the threshold will be masked by 'N'.
Sequences at the edges of the sequences will also be masked by 'N'.
}
\examples{
res <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1])
new_sequences <- mask_sequences_by_pwm(cluster_sequences_example, res$pssm, quantile(res$pred, 0.95), spat = res$spat)
head(new_sequences)

}
