% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-dbs.R
\docType{data}
\name{MOTIF_DB}
\alias{MOTIF_DB}
\title{A MotifDB object with motifs from all bundled databases}
\format{
A MotifDB object
}
\usage{
MOTIF_DB
}
\description{
A MotifDB object with motifs from all bundled databases
}
\keyword{datasets}
