% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{motif_db_to_dataframe}
\alias{motif_db_to_dataframe}
\title{Convert a MotifDB object back to a tidy data frame}
\usage{
motif_db_to_dataframe(motif_db)
}
\arguments{
\item{motif_db}{A MotifDB object}
}
\value{
A tidy data frame with columns for motif, position, and nucleotide probabilities
}
\description{
Convert a MotifDB object back to a tidy data frame
}
\examples{
head(motif_db_to_dataframe(MOTIF_DB))
}
