% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\docType{class}
\name{MotifDB-class}
\alias{MotifDB-class}
\title{MotifDB Class}
\description{
S4 class to store position weight matrices and their properties
}
\section{Slots}{

\describe{
\item{\code{mat}}{A numeric matrix containing position weight matrices in log scale}

\item{\code{rc_mat}}{A numeric matrix containing reverse complement of position weight matrices}

\item{\code{motif_lengths}}{A named integer vector containing the length of each motif}

\item{\code{prior}}{The pssm prior probability}

\item{\code{spat_factors}}{A numeric matrix containing spatial factors}

\item{\code{spat_bin_size}}{The size of spatial bins}

\item{\code{spat_min}}{The starting position of the sequence or NA}

\item{\code{spat_max}}{The ending position of the sequence or NA}
}}

