% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{names,MotifDB-method}
\alias{names,MotifDB-method}
\title{Get the names of motifs in a MotifDB object}
\usage{
\S4method{names}{MotifDB}(x)
}
\arguments{
\item{x}{MotifDB object}
}
\value{
The names of motifs in the object
}
\description{
Get the names of motifs in a MotifDB object
}
\examples{
names(MOTIF_DB)
names(MOTIF_DB[c("HOMER.GATA3_2", "JASPAR.CDX1")])
}
