% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{plot,MotifDB-method}
\alias{plot,MotifDB-method}
\title{Plot a motif from a MotifDB object}
\usage{
\S4method{plot}{MotifDB}(
  x,
  title = "",
  subtitle = ggplot2::waiver(),
  revcomp = FALSE,
  method = "bits",
  force = FALSE,
  ...
)
}
\arguments{
\item{x}{MotifDB object}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{revcomp}{whether to plot the reverse complement of the PSSM}

\item{method}{Height method, can be one of "bits" or "probability" (default:"bits")}

\item{force}{force plotting more than 30 motifs}

\item{...}{additional arguments passed to \code{ggseqlogo::ggseqlogo()}}
}
\value{
a ggplot object
}
\description{
Plot a motif from a MotifDB object
}
\examples{
plot(MOTIF_DB["HOMER.GATA3_2"])
plot(MOTIF_DB["HNF1", pattern = TRUE])
plot(MOTIF_DB[c("HOMER.GATA3_2", "JASPAR.CDX1")])

}
