% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-regression.R
\name{plot_spat_model}
\alias{plot_spat_model}
\title{Plot spatial model of the regression result}
\usage{
plot_spat_model(spat, title = "Spatial model")
}
\arguments{
\item{spat}{the 'spat' field from the regression result}

\item{title}{a title for the plot (optional)}
}
\description{
Plot spatial model of the regression result
}
\examples{
\dontrun{
res <- regress_pwm(sequences_example, response_mat_example)
plot_spat_model(res$spat)
}

}
