% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prego-package.R
\docType{package}
\name{prego-package}
\alias{prego}
\alias{prego-package}
\title{prego: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

Authors:
\itemize{
  \item Amos Tanay \email{amos.tanay@weizmann.ac.il}
}

}
\keyword{internal}
