% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{prior,MotifDB-method}
\alias{prior,MotifDB-method}
\title{Get the prior value from a MotifDB object}
\usage{
\S4method{prior}{MotifDB}(object)
}
\arguments{
\item{object}{MotifDB object}
}
\value{
The prior value
}
\description{
Get the prior value from a MotifDB object
}
