% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{prior<-,MotifDB-method}
\alias{prior<-,MotifDB-method}
\title{Set a new prior for a MotifDB object}
\usage{
\S4method{prior}{MotifDB}(object) <- value
}
\arguments{
\item{object}{MotifDB object}

\item{value}{New prior value between 0 and 1}
}
\value{
Updated MotifDB object with new prior
}
\description{
Set a new prior for a MotifDB object
}
\examples{
prior(MOTIF_DB)
prior(MOTIF_DB) <- 0.2
prior(MOTIF_DB)
}
