% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{prior<-}
\alias{prior<-}
\title{Set the prior for a MotifDB object}
\usage{
prior(object) <- value
}
\arguments{
\item{object}{MotifDB object}

\item{value}{New prior value}
}
\value{
Updated MotifDB object
}
\description{
Set the prior for a MotifDB object
}
