% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{prior}
\alias{prior}
\title{Get or set the prior for a MotifDB object}
\usage{
prior(object)
}
\arguments{
\item{object}{MotifDB object}
}
\value{
The prior value
}
\description{
Get or set the prior for a MotifDB object
}
