% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-cor.R
\name{pssm_cor}
\alias{pssm_cor}
\title{Compute the correlation between two given PSSMs}
\usage{
pssm_cor(pssm1, pssm2, method = c("spearman", "pearson"), prior = 0.01)
}
\arguments{
\item{pssm1}{first PSSM matrix or data frame}

\item{pssm2}{second PSSM matrix or data frame}

\item{method}{method to use for computing the correlation. See \code{\link[stats]{cor}} for details.}

\item{prior}{a prior probability for each nucleotide.}
}
\value{
Correlation between the two PSSMs
}
\description{
The correlation is computed by shifting the shorter PSSM along the longer one
and computing the correlation at each position. The maximum correlation is returned.
}
\examples{
\dontrun{
res1 <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1])
pssm_cor(res1$pssm, JASPAR_motifs[JASPAR_motifs$motif == "HNF1A", ])
}

# Compare two motifs from MOTIF_DB
pssm_cor(as.matrix(MOTIF_DB["HOMER.GATA3_2"]), as.matrix(MOTIF_DB["JASPAR.CDX1"]))

# Compare using different correlation methods
pssm_cor(as.matrix(MOTIF_DB["HOMER.GATA3_2"]), as.matrix(MOTIF_DB["JASPAR.GATA3"]), method = "pearson")
pssm_cor(as.matrix(MOTIF_DB["HOMER.GATA3_2"]), as.matrix(MOTIF_DB["JASPAR.GATA3"]), method = "spearman")

}
