% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{pssm_dataset_cor}
\alias{pssm_dataset_cor}
\title{Compute a correlation matrix for a pssm dataset}
\usage{
pssm_dataset_cor(
  dataset,
  method = "spearman",
  prior = 0.01,
  parallel = getOption("prego.parallel", TRUE)
)
}
\arguments{
\item{dataset}{a pssm dataset. A data frame with the columns 'motif', 'pos', 'A", 'C', 'G', 'T'}

\item{method}{method to use for computing the correlation. See \code{\link[stats]{cor}} for details.}

\item{prior}{a prior probability for each nucleotide.}

\item{parallel}{whether to use parallel computing}
}
\value{
A correlation matrix for the PSSMs in the dataset
}
\description{
Compute a correlation matrix for a pssm dataset
}
\examples{
\dontrun{
cm <- pssm_dataset_cor(JASPAM_motifs)
head(cm)
}

}
