% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-cor.R
\name{pssm_diff}
\alias{pssm_diff}
\title{Compute the KL divergence between two given PSSMs}
\usage{
pssm_diff(pssm1, pssm2, prior = 0.01)
}
\arguments{
\item{pssm1}{first PSSM matrix or data frame}

\item{pssm2}{second PSSM matrix or data frame}

\item{prior}{a prior probability for each nucleotide.}
}
\value{
KL divergence between the two PSSMs (lower values indicate more similar PSSMs)
}
\description{
The KL divergence is computed by shifting the shorter PSSM along the longer one
and computing the divergence at each position. The minimum divergence is returned.
}
\examples{
\dontrun{
res1 <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1])
pssm_diff(res1$pssm, JASPAR_motifs[JASPAR_motifs$motif == "HNF1A", ])
}

}
