% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer-regression.R
\name{pssm_to_kmer}
\alias{pssm_to_kmer}
\title{Transform PSSM (Position-Specific Scoring Matrix) to a KMER}
\usage{
pssm_to_kmer(pssm, kmer_length = 7, pos_bits_thresh = NULL)
}
\arguments{
\item{pssm}{PSSM matrix or data frame. The PSSM must have at least kmer_length rows.}

\item{kmer_length}{The length of the k-mer to return.}

\item{pos_bits_thresh}{A numeric value indicating the minimum number of bits per position to include the nucleotide in the k-mer. If the nucleotide does not meet this threshold, it is replaced with 'N'. Default is NULL.}
}
\value{
A character vector of length 1 containing the k-mer.
}
\description{
This function transforms a PSSM into a k-mer of a given length.
}
\examples{
pssm_to_kmer(get_motif_pssm("HOMER.AP_1"))
plot_pssm_logo_dataset("HOMER.AP_1")

}
