% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sample.R
\name{sample_quantile_matched_rows}
\alias{sample_quantile_matched_rows}
\title{Sample rows respecting quantiles of a reference distribution}
\usage{
sample_quantile_matched_rows(
  data_frame,
  reference,
  sample_fraction,
  num_quantiles = 10,
  seed = 60427
)
}
\arguments{
\item{data_frame}{A data frame from which to sample rows.}

\item{reference}{A numeric vector of the same length as the number of rows in the data frame.}

\item{sample_fraction}{A fraction specifying the proportion of rows to sample from the data frame.}

\item{num_quantiles}{An integer specifying the number of quantiles, default is 10.}

\item{seed}{An integer specifying the random seed to use.}
}
\value{
A data frame of sampled rows.
}
\description{
This function randomly samples rows from a data frame in such a way that the
quantiles of the selected data match as closely as possible those of the full data.
}
\examples{
sampled <- sample_quantile_matched_rows(mtcars, mtcars$mpg, sample_fraction = 0.1)
plot(quantile(mtcars$mpg), quantile(sampled$mpg))
abline(0, 1)

}
