% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{screen_local_pwm}
\alias{screen_local_pwm}
\title{Screen sequences for positions with PWM scores meeting a threshold condition}
\usage{
screen_local_pwm(
  sequences,
  pssm,
  operator,
  threshold,
  spat = NULL,
  spat_min = 0,
  spat_max = NULL,
  bidirect = TRUE,
  prior = 0.01
)
}
\arguments{
\item{sequences}{a vector of sequences}

\item{pssm}{a PSSM matrix or data frame. The columns of the matrix or data frame should be named with the nucleotides ('A', 'C', 'G' and 'T').}

\item{operator}{A character string specifying the comparison operator. One of: ">", "<", ">=", "<=", "=="}

\item{threshold}{A numeric value specifying the threshold for comparison}

\item{spat}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}

\item{spat_min}{the minimum position to use from the sequences. The default is 1.}

\item{spat_max}{the maximum position to use from the sequences. The default is the length of the sequences.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{prior}{a prior probability for each nucleotide.}
}
\value{
A list with one element per sequence, where each element is a numeric vector
containing the 1-indexed positions that meet the threshold condition.
}
\description{
This function screens sequences to find positions where the local PWM score
meets a specified threshold condition using various operators (>, <, >=, <=, ==).
}
\examples{
\dontrun{
# Find positions where HNF1A motif scores above -19
hnf1a_positions <- screen_local_pwm(
    cluster_sequences_example,
    as.matrix(MOTIF_DB["JASPAR.HNF1A"]),
    operator = ">",
    threshold = -37
)

which(purrr::map_dbl(hnf1a_positions, length) > 0) # which sequences have positions?
hnf1a_positions[[4]] # positions for the 4th sequence

hnf1a_energy <- compute_local_pwm(cluster_sequences_example, as.matrix(MOTIF_DB["JASPAR.HNF1A"]))
hnf1a_energy[4, hnf1a_positions[[4]]]
}

}
