% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-clusters.R
\name{screen_pwm.clusters}
\alias{screen_pwm.clusters}
\title{Screen for motifs in a database for every cluster}
\usage{
screen_pwm.clusters(
  sequences,
  clusters,
  dataset = all_motif_datasets(),
  motifs = NULL,
  parallel = getOption("prego.parallel", TRUE),
  min_D = 0.4,
  only_best = FALSE,
  prior = 0.01,
  alternative = "two.sided",
  ...
)
}
\arguments{
\item{sequences}{a vector with the sequences}

\item{clusters}{a vector with the cluster assignments}

\item{dataset}{a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example \code{HOMER_motifs} or \code{JASPAR_motifs}, or \code{all_motif_datasets()}.}

\item{motifs}{names of specific motifs to extract from the dataset}

\item{parallel}{logical, whether to use parallel processing}

\item{min_D}{minimum distance to consider a match}

\item{only_best}{if TRUE, only return the best match for each cluster}

\item{prior}{a prior probability for each nucleotide.}

\item{alternative}{alternative hypothesis for the KS test. Can be "two.sided", "less" or "greater"}

\item{...}{
  Arguments passed on to \code{\link[=compute_pwm]{compute_pwm}}
  \describe{
    \item{\code{pssm}}{a PSSM matrix or data frame. The columns of the matrix or data frame should be named with the nucleotides ('A', 'C', 'G' and 'T').}
    \item{\code{spat}}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}
    \item{\code{bidirect}}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}
    \item{\code{spat_min}}{start of the spatial model from the beginning of the sequence (in bp)}
    \item{\code{spat_max}}{end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model
would end at the end of the sequence.}
  }}
}
\value{
a matrix with the KS D statistics for each cluster (columns) and every motif (rows)
that had at least one cluster with D >= min_D. If \code{only_best} is TRUE, a named vector
with the name of best motif match for each cluster is returned (regardless of \code{min_D}).
}
\description{
Screen for motifs in a database for every cluster
}
\examples{
\dontrun{
D_mat <- screen_pwm.clusters(cluster_sequences_example, clusters_example)
dim(D_mat)
D_mat[1:5, 1:5]

# return only the best match
screen_pwm.clusters(cluster_sequences_example, clusters_example, only_best = TRUE)
}

}
