% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen-pwm.R
\name{screen_pwm}
\alias{screen_pwm}
\title{Screen for motifs in a database given a response variable}
\usage{
screen_pwm(
  sequences,
  response,
  metric = NULL,
  dataset = all_motif_datasets(),
  motifs = NULL,
  parallel = getOption("prego.parallel", TRUE),
  only_best = FALSE,
  prior = 0.01,
  alternative = "two.sided",
  ...
)
}
\arguments{
\item{sequences}{a vector with the sequences}

\item{response}{a vector of response variable for each sequence. If the response is a matrix, the average will be used.}

\item{metric}{metric to use in order to choose the best motif. One of 'ks' or 'r2'. If NULL - the default would be 'ks' for binary variables, and 'r2' for continuous variables.}

\item{dataset}{a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example \code{HOMER_motifs} or \code{JASPAR_motifs}, or \code{all_motif_datasets()}.}

\item{motifs}{names of specific motifs to extract from the dataset}

\item{parallel}{logical, whether to use parallel processing}

\item{only_best}{return only the best motif (the one with the highest score). If FALSE, all the motifs will be returned.}

\item{prior}{a prior probability for each nucleotide.}

\item{alternative}{alternative hypothesis for the KS test. One of 'two.sided', 'less' or 'greater'.}

\item{...}{
  Arguments passed on to \code{\link[=compute_pwm]{compute_pwm}}
  \describe{
    \item{\code{pssm}}{a PSSM matrix or data frame. The columns of the matrix or data frame should be named with the nucleotides ('A', 'C', 'G' and 'T').}
    \item{\code{spat}}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}
    \item{\code{spat_min}}{the minimum position to use from the sequences. The default is 1.}
    \item{\code{spat_max}}{the maximum position to use from the sequences. The default is the length of the sequences.}
    \item{\code{bidirect}}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}
    \item{\code{func}}{the function to use to combine the PWMs for each sequence. Either 'logSumExp' or 'max'. The default is 'logSumExp'.}
  }}
}
\value{
a data frame with the following columns:
\describe{
\item{motif: }{the motif name.}
\item{score: }{the score of the motif (depending on \code{metric}).}
}

if \code{only_best} is TRUE, only the best motif would be returned (a data framw with a single row).
}
\description{
Screen for motifs in a database given a response variable
}
\examples{
res_screen <- screen_pwm(cluster_sequences_example, cluster_mat_example[, 1])
head(res_screen)

# only best match
screen_pwm(cluster_sequences_example, cluster_mat_example[, 1])

# with r^2 metric
res_screen <- screen_pwm(sequences_example, response_mat_example[, 1], metric = "r2")
head(res_screen)

}
