% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_parallel}
\alias{set_parallel}
\title{Set parallel threads}
\usage{
set_parallel(thread_num = max(1, round(parallel::detectCores() * 0.8)))
}
\arguments{
\item{thread_num}{number of threads. use '1' for non parallel behavior}
}
\value{
None
}
\description{
Set the number of parallel threads to use. prego uses the R function \code{doMC::registerDoMC} to register the parallelization.
By default, prego uses 80\% of the number of available cores. The options are saved under 'prego.parallel' (should we use parallelization, logical) and 'prego.parallel.nc' (number of cores to use, integer).
}
\examples{
\donttest{
set_parallel(8)
}
}
