% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{show,MotifDB-method}
\alias{show,MotifDB-method}
\title{Show method for MotifDB objects}
\usage{
\S4method{show}{MotifDB}(object)
}
\arguments{
\item{object}{MotifDB object}
}
\description{
Show method for MotifDB objects
}
