% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{[,MotifDB-method}
\alias{[,MotifDB-method}
\title{Get specific motifs from the MotifDB}
\usage{
\S4method{[}{MotifDB}(x, i)
}
\arguments{
\item{object}{MotifDB object}

\item{motif_names}{Character vector of motif names or numeric indices}
}
\value{
MotifDB object containing the specified motifs
}
\description{
Get specific motifs from the MotifDB
}
