% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-dbs.R
\name{all_motif_datasets}
\alias{all_motif_datasets}
\title{Get a data frame of all the motif datasets bundled with prego}
\usage{
all_motif_datasets()
}
\value{
a data frame which concatenates motifs from "HOMER", "JASPAR" and "JOLMA".
Motif names are prefixed with the dataset name, e.g. "JASPAR.GATA4".
}
\description{
The data frame contain the PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name. Individual datasets are available within the package as \code{HOMER_motifs}, \code{JASPAR_motifs}, and \code{JOLMA_motifs}.
}
\examples{
all_motif_datasets()

}
