% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{compute_pwm}
\alias{compute_pwm}
\title{Compute PWMs for a set of sequences given a PSSM matrix}
\usage{
compute_pwm(
  sequences,
  pssm,
  spat = NULL,
  spat_min = 1,
  spat_max = NULL,
  bidirect = TRUE,
  prior = 0,
  func = "logSumExp"
)
}
\arguments{
\item{sequences}{a vector of sequences}

\item{pssm}{a PSSM matrix or data frame. The columns of the matrix or data frame should be named with the nucleotides ('A', 'C', 'G' and 'T').}

\item{spat}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}

\item{spat_min}{the minimum position to use from the sequences. The default is 1.}

\item{spat_max}{the maximum position to use from the sequences. The default is the length of the sequences.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{prior}{a prior probability for each nucleotide.}

\item{func}{the function to use to combine the PWMs for each sequence. Either 'logSumExp' or 'max'. The default is 'logSumExp'.}
}
\value{
a vector with the predicted pwm for each sequence.
}
\description{
Compute PWMs for a set of sequences given a PSSM matrix
}
\examples{
\dontrun{
res <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1])

pwm <- compute_pwm(cluster_sequences_example, res$pssm, res$spat)
head(pwm)

# this is similar to the prediction in the regression
head(res$pred)
}

}
