% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{create_motif_db}
\alias{create_motif_db}
\title{Create a MotifDB object from a tidy data frame}
\usage{
create_motif_db(
  motif_db,
  prior = 0.01,
  spat_factors = NULL,
  spat_bin_size = 1,
  spat_min = NA_real_,
  spat_max = NA_real_
)
}
\arguments{
\item{motif_db}{A tidy data frame containing motif information}

\item{prior}{Pseudocount prior to add to probabilities (default: 0.01)}

\item{spat_factors}{Matrix of spatial factors (rows=motifs, cols=bins) or NULL}

\item{spat_bin_size}{Size of spatial bins (default: 1)}

\item{spat_min}{Starting position of sequence or NA (default: NA)}

\item{spat_max}{Ending position of sequence or NA (default: NA)}
}
\value{
A MotifDB object
}
\description{
Create a MotifDB object from a tidy data frame
}
\examples{
create_motif_db(all_motif_datasets())
}
