% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif-dbs.R
\name{gextract_pwm}
\alias{gextract_pwm}
\title{Extract pwm of intervals from a motif database}
\usage{
gextract_pwm(
  intervals,
  motifs = NULL,
  dataset = all_motif_datasets(),
  spat = NULL,
  spat_min = 0,
  spat_max = NULL,
  bidirect = TRUE,
  parallel = getOption("prego.parallel", TRUE)
)
}
\arguments{
\item{intervals}{misha intervals set}

\item{motifs}{names of specific motifs to extract from the dataset}

\item{dataset}{a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example \code{HOMER_motifs} or \code{JASPAR_motifs}, or \code{all_motif_datasets()}.}

\item{spat}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}

\item{spat_min}{start of the spatial model from the beginning of the sequence (in bp)}

\item{spat_max}{end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model
would end at the end of the sequence.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{parallel}{logical, whether to use parallel processing}
}
\value{
The intervals set with additional columns per motif, containing the pwm of each interval for each motif
}
\description{
Extract pwm of intervals from a motif database
}
\examples{
\dontrun{
library(misha)
gdb.init_examples()
pwms <- gextract_pwm("annotations")
pwms[, 1:20]
}

}
