% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misha.R
\name{gpwm_quantiles}
\alias{gpwm_quantiles}
\title{Compute quantile of pwm for a given interval size}
\usage{
gpwm_quantiles(
  size,
  quantiles,
  pssm,
  spat = NULL,
  spat_min = 1,
  spat_max = NULL,
  bidirect = TRUE,
  prior = 0.01,
  n_sequences = 10000,
  dist_from_edge = 3000000,
  chromosomes = NULL,
  func = "logSumExp"
)
}
\arguments{
\item{size}{size of the intervals to sample}

\item{quantiles}{quantiles to compute. See \code{quantile} for more details.}

\item{pssm}{PSSM matrix or data frame}

\item{spat}{a data frame with the spatial model (as returned from the \code{$spat} slot from the regression). Should contain a column called 'bin' and a column called 'spat_factor'.}

\item{spat_min}{the minimum position to use from the sequences. The default is 1.}

\item{spat_max}{the maximum position to use from the sequences. The default is the length of the sequences.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{prior}{a prior probability for each nucleotide.}

\item{n_sequences}{number of sequences to sample in order to compute the quantiles. The default is 1e4.}

\item{dist_from_edge}{The minimum distance from the edge of the chromosome for a region to start or end(default: 3e6)}

\item{chromosomes}{The chromosomes to sample from (default: all chromosomes)}

\item{func}{the function to use to combine the PWMs for each sequence. Either 'logSumExp' or 'max'. The default is 'logSumExp'.}
}
\value{
a named vector with the quantiles of the pwm for the given interval size.
}
\description{
Computes the quantile of the pwm for a given interval size by sampling random intervals from the genome. The number of sequences to sample can be specified with \code{n_sequences}.
}
\examples{
\dontrun{
library(misha)
library(dplyr)
gdb.init_examples()
pssm <- JASPAR_motifs \%>\%
    filter(motif == "JASPAR.CDX1") \%>\%
    select(-motif)
gpwm_quantiles(1000, seq(0, 1, 0.1), pssm, dist_from_edge = 100)
}

}
