% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{initialize,MotifDB-method}
\alias{initialize,MotifDB-method}
\title{Initialize method for MotifDB objects}
\usage{
\S4method{initialize}{MotifDB}(
  .Object,
  mat = matrix(0, 4, 1),
  rc_mat = matrix(0, 4, 1),
  motif_lengths = setNames(as.integer(1), colnames(mat)[1]),
  prior = 0.01,
  spat_factors = matrix(1, nrow = 1, ncol = 1),
  spat_bin_size = 1
)
}
\arguments{
\item{.Object}{MotifDB object}

\item{mat}{Position weight matrix}

\item{rc_mat}{Reverse complement matrix}

\item{motif_lengths}{Named vector of motif lengths}

\item{prior}{Pseudocount prior value}

\item{spat_factors}{Matrix of spatial factors}

\item{spat_bin_size}{Size of spatial bins}
}
\description{
Initialize method for MotifDB objects
}
