% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer-regression.R
\name{kmer_matrix}
\alias{kmer_matrix}
\title{Generate a kmer Matrix}
\usage{
kmer_matrix(
  sequences,
  kmer_length,
  from_range = 1,
  to_range = NULL,
  set_rownames = FALSE
)
}
\arguments{
\item{sequences}{A vector of strings with DNA sequences ('T', 'C', 'G', 'A' or 'N').}

\item{from_range}{Sequences will be considered only from position from_range.}

\item{to_range}{Sequences will be considered only up to position to_range (default NULL - using the length of the sequences).}

\item{set_rownames}{If TRUE, the rownames of the matrix will be set to the sequences (default FALSE).}
}
\value{
A matrix where rows are the number of sequences, columns are the kmers and the values are the number of occurrences of each kmer.
}
\description{
This function calculates the frequency of each kmer for each DNA sequence.
}
\examples{
kmer_matrix(c("ATCG", "TCGA", "ATAT"), 2)
kmer_matrix(c("ATCG", "TCGA", "ATAT"), 3)

}
