% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{load_multi_regression}
\alias{load_multi_regression}
\title{Load a multiple motif regression model}
\usage{
load_multi_regression(fn)
}
\arguments{
\item{fn}{file name}
}
\value{
a list with the following elements:

\itemize{
\item{models: }{a list of models.}
\item{model: }{the combined model.}
\item{spat_min: }{the minimum spatial position.}
\item{spat_max: }{the maximum spatial position.}
\item{bidirect: }{whether the model is bidirectional.}
\item{spat_bin_size: }{the spatial bin size.}
\item{seq_length: }{the sequence length.}
\item{motif_num: }{the number of motifs.}
\item{predict: }{a function to predict the response.}
\item{predict_multi: }{a function to predict the response for each motif.}
}
}
\description{
Load a multiple motif regression model
}
\examples{
\dontrun{
res_multi <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1],
    final_metric = "ks", spat_bin_size = 40,
    spat_num_bins = 7,
    motif_num = 2
)
tmp <- tempfile()
res_multi$export(tmp)
r <- load_multi_regression(tmp)
}

}
