% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifDB.R
\name{plot,MotifDB-method}
\alias{plot,MotifDB-method}
\title{Plot a motif from a MotifDB object}
\usage{
\S4method{plot}{MotifDB}(
  x,
  title = "",
  subtitle = ggplot2::waiver(),
  revcomp = FALSE,
  method = "bits",
  ...
)
}
\arguments{
\item{x}{MotifDB object}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{revcomp}{whether to plot the reverse complement of the PSSM}

\item{method}{Height method, can be one of "bits" or "probability" (default:"bits")}

\item{...}{
  Arguments passed on to \code{\link[ggseqlogo:ggseqlogo]{ggseqlogo::ggseqlogo}}
  \describe{
    \item{\code{data}}{Character vector of sequences or named list of sequences. All sequences must have same width}
    \item{\code{facet}}{Facet type, can be 'wrap' or 'grid'}
    \item{\code{scales}}{Facet scales, see \code{\link{facet_wrap}}}
    \item{\code{ncol}}{Number of columns, works only when \code{facet='wrap'}, see \code{\link{facet_wrap}}}
    \item{\code{nrow}}{Number of rows, same as \code{ncol}}
  }}
}
\value{
a ggplot object
}
\description{
Plot a motif from a MotifDB object
}
\examples{
plot(MOTIF_DB["HOMER.GATA3_2"])
plot(MOTIF_DB["HNF1", pattern = T])
plot(MOTIF_DB[c("HOMER.GATA3_2", "JASPAR.CDX1")])

}
