% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{plot_pssm_logo}
\alias{plot_pssm_logo}
\title{Plot LOGO of the pssm result from the regression}
\usage{
plot_pssm_logo(pssm, title = "Sequence model", subtitle = ggplot2::waiver())
}
\arguments{
\item{pssm}{the 'pssm' field from the regression result}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}
}
\value{
a ggplot object
}
\description{
Plot LOGO of the pssm result from the regression
}
\examples{
\dontrun{
res <- regress_pwm(sequences_example, response_mat_example)
plot_pssm_logo(res$pssm)
}

}
