% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-regression.R
\name{plot_pssm_logo}
\alias{plot_pssm_logo}
\title{Plot LOGO of the pssm result from the regression}
\usage{
plot_pssm_logo(pssm)
}
\arguments{
\item{pssm}{the 'pssm' field from the regression result}
}
\description{
Plot LOGO of the pssm result from the regression
}
\examples{
res <- regress_pwm(sequences_example, response_mat_example)
plot_pssm_logo(res$pssm)

}
