% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{plot_pssm_logo}
\alias{plot_pssm_logo}
\title{Plot LOGO of the pssm result from the regression}
\usage{
plot_pssm_logo(
  pssm,
  title = "Sequence model",
  subtitle = ggplot2::waiver(),
  pos_bits_thresh = NULL
)
}
\arguments{
\item{pssm}{PSSM matrix or data frame}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{pos_bits_thresh}{Positions with bits above this threshold would be highlighted in red. If \code{NULL}, no positions would be highlighted.}
}
\value{
a ggplot object
}
\description{
Plot LOGO of the pssm result from the regression
}
\examples{
\dontrun{
res <- regress_pwm(sequences_example, response_mat_example)
plot_pssm_logo(res$pssm)
}

}
