% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{plot_regression_qc}
\alias{plot_regression_qc}
\title{Plot the regression results}
\usage{
plot_regression_qc(reg)
}
\arguments{
\item{reg}{output of \code{regress_pwm}}
}
\value{
a patchwork object
}
\description{
Plot QC of the regression results
}
\examples{
res <- regress_pwm(sequences_example, response_mat_example)
plot_regression_qc(res)

}
