% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-regression.R
\name{plot_regression_qc_multi}
\alias{plot_regression_qc_multi}
\title{Plot the regression results for multiple motifs}
\usage{
plot_regression_qc_multi(
  reg,
  title = glue("Motif regression results (consensus: {reg$consensus})"),
  subtitle = NULL,
 
    caption = glue("# of 1: {sum(response == 1)}, # of 0: {sum(response == 0)}, seed: {reg$seed_motif}")
)
}
\arguments{
\item{reg}{output of \code{regress_pwm}}

\item{title}{a title for the plot (optional)}

\item{subtitle}{a subtitle for the plot (optional)}

\item{caption}{a caption for the plot (optional)}
}
\description{
plot the regression results when \code{motif_num} > 1
}
\examples{
res_binary <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 3], motif_num = 3)
plot_regression_qc_multi(res_binary)

}
