% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prego-package.R
\docType{package}
\name{prego-package}
\alias{prego}
\alias{prego-package}
\title{prego: PWM Regression Optimizer}
\description{
Implements regression algorithms for finding motifs in DNA. It contains algorithms which were described at: "Extensive low-affinity transcriptional interactions in the yeast genome" \doi{10.1101/gr.5113606}.
}
\author{
\strong{Maintainer}: Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}

Authors:
\itemize{
  \item Amos Tanay \email{amos.tanay@weizmann.ac.il}
}

}
\keyword{internal}
