% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{pssm_cor}
\alias{pssm_cor}
\title{Compute the correlation between two given PSSMs}
\usage{
pssm_cor(pssm1, pssm2)
}
\arguments{
\item{pssm1}{first PSSM matrix or data frame}

\item{pssm2}{second PSSM matrix or data frame}
}
\value{
Correlation between the two PSSMs
}
\description{
The correlation is computed by shifting the shorter PSSM along the longer one
and computing the correlation at each position. The maximum correlation is returned.
}
\examples{
\dontrun{
res1 <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1])
pssm_cor(res1$pssm, JASPAR_motifs[JASPAR_motifs$motif == "HNF1A", ])
}

}
