% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pssm-utils.R
\name{pssm_diff}
\alias{pssm_diff}
\title{Compute KL divergence between two PSSMs}
\usage{
pssm_diff(pssm1, pssm2)
}
\arguments{
\item{pssm1}{first PSSM matrix or data frame}

\item{pssm2}{second PSSM matrix or data frame}
}
\value{
KL divergence between the two PSSMs
}
\description{
Compute KL divergence between two PSSMs
}
\examples{
\dontrun{
res1 <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1])
pssm_diff(res1$pssm, JASPAR_motifs[JASPAR_motifs$motif == "HNF1A", ])
}

}
