% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regress_pwm}
\alias{regress_pwm}
\title{Perform a PWM regression}
\usage{
regress_pwm(
  sequences,
  response,
  motif = NULL,
  motif_length = 15,
  score_metric = "r2",
  bidirect = TRUE,
  spat_min = 0,
  spat_max = NULL,
  spat_bin = 50,
  spat_model = NULL,
  improve_epsilon = 0.0001,
  min_nuc_prob = 0.001,
  unif_prior = 0.05,
  is_train = NULL,
  include_response = TRUE,
  seed = 60427,
  verbose = FALSE,
  kmer_length = 8,
  motif_num = 1,
  consensus_single_thresh = 0.6,
  consensus_double_thresh = 0.85,
  ...
)
}
\arguments{
\item{sequences}{A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through \code{toupper})}

\item{response}{A matrix of response variables - number of rows should equal the number of sequences}

\item{motif}{Initial motif to start the regression from. Can be either a string with a kmer where the character "*" indicates a
wildcard or a data frame with a pre-computed PSSM (see thre slot \code{pssm} in the return value of this function).
If NULL - a K-mer screen would be performed in order to find the best kmer for initialization.}

\item{motif_length}{Length of the seed motif. If the motif is shorter than this, it will be extended by wildcards (stars). Note that If the motif is longer than this, it will \emph{not} be truncated.}

\item{score_metric}{metric to use for optimizing the PWM. One of "r2" or "ks". When using "ks" the response variable should be a single vector of 0 and 1.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{spat_min}{start of the spatial model from the beginning of the sequence (in bp)}

\item{spat_max}{end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model
would end at the end of the sequence.}

\item{spat_bin}{size of the spatial bin (in bp).}

\item{spat_model}{a previously computed spatial model (see \code{spat}) in the return value of this function. This can only be used when \code{motif} is a previously computed PSSM.}

\item{improve_epsilon}{minimum improve in the objective function to continue the optimization}

\item{min_nuc_prob}{minimum nucleotide probability in every iteration}

\item{unif_prior}{uniform prior for nucleotide probabilities}

\item{is_train}{a boolean vector that determine which subset of sequences to use when screening}

\item{include_response}{include the response in the resulting list (default: TRUE)}

\item{seed}{random seed}

\item{verbose}{show verbose messages.}

\item{kmer_length}{The number of non-gap characters in motifs that will be screened}

\item{consensus_single_thresh, consensus_double_thresh}{thresholds for the consensus sequence calculation
(single and double nucleotides)}

\item{...}{
  Arguments passed on to \code{\link[=screen_kmers]{screen_kmers}}
  \describe{
    \item{\code{min_cor}}{Only patterns for which the maximum correlation to one of the response variable is larger than min_cor will be reported}
    \item{\code{min_n}}{Only patterns for which the average number of occurrences in the sequences is larger than min_n will be reported}
    \item{\code{min_gap,max_gap}}{the length of a gap to be considered in the pattern. Only one gap, of length min_gap:max_gap, is being used, and is located anywhere in the motif. Note that this greatly expand the search space (and increase multiple tesing severly).}
    \item{\code{from_range}}{Sequences will be considered only from position from_range (default 0)}
    \item{\code{to_range}}{Sequences will be considered only up to position to_range (default NULL - using the length of the sequences)}
    \item{\code{return_mat}}{Return a matrix of patterns and their correlation to the response variables instead of a data frame. (default: FALSE)}
  }}
}
\value{
a list with the following elements:
\itemize{
\item{pssm: }{data frame with the pssm matrix with the inferred motif, where rows are positions and columns are nucleotides.}
\item{spat: }{a data frame with the inferred spatial model, with the spatial factor for each bin.}
\item{pred: }{a vector with the predicted pwm for each sequence.}
\item{consensus: }{Consensus sequence based on the PSSM.}
\item{response: }{The response matrix. If \code{include_response} is FALSE, the response matrix is not included in the list.}
\item{r2: }{\eqn{r^2} of the prediction with respect to the each response variable.}
\item{ks: }{If response is binary, Kolmogorov-Smirnov test results of the predictions where the response was 1 vs the predictions where the response was 0.}
\item{seed_motif: }{The seed motif that started the regression.}
\item{kmers: }{The k-mers that were screened in order to find the best seed motif (if motif was NULL).}
}
}
\description{
Perform a PWM regression
}
\examples{
res <- regress_pwm(sequences_example, response_mat_example)
res$pssm
res$spat
head(res$pred)

plot_regression_qc(res)

# intialize with a pre-computed PSSM
res1 <- regress_pwm(sequences_example, response_mat_example, motif = res$pssm)

# intialize with a pre-computed PSSM and spatial model
res2 <- regress_pwm(sequences_example, response_mat_example, motif = res$pssm, spat_model = res$spat)

# binary response
res_binary <- regress_pwm(cluster_sequences_example, cluster_mat_example[, 1])
plot_regression_qc(res_binary)

}
