% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regress_pwm}
\alias{regress_pwm}
\title{Perform a PWM regression}
\usage{
regress_pwm(
  sequences,
  response,
  motif = NULL,
  motif_length = 8,
  score_metric = "r2",
  bidirect = TRUE,
  spat_min = 0,
  spat_max = NULL,
  spat_bin = 50,
  improve_epsilon = 0.0001,
  min_nuc_prob = 0.001,
  unif_prior = 0.05,
  is_train = NULL,
  include_response = TRUE,
  seed = 60427,
  verbose = FALSE,
  kmer_length = 8,
  motif_num = 1,
  ...
)
}
\arguments{
\item{sequences}{A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through \code{toupper})}

\item{response}{A matrix of response variables - number of rows should equal the number of sequences}

\item{motif}{Initial motif to start the regression from. The character "*" indicates a wildcard.
If NULL - a K-mer screen would be performed in order to find the best kmer for initialization.}

\item{motif_length}{Length of the seed motif. If the motif is shorter than this, it will be extended by wildcards (stars). Note that If the motif is longer than this, it will \emph{not} be truncated.}

\item{score_metric}{metric to use for optimizing the PWM. One of "r2" or "ks". For categorical response variables (0 and 1), "ks" is recommended, while for continuous response variables, "r2" is recommended. Default is "r2". When using "ks" the response variable should be a single vector of 0 and 1.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{spat_min}{start of the spatial model from the beginning of the sequence (in bp)}

\item{spat_max}{end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model
would end at the end of the sequence.}

\item{spat_bin}{size of the spatial bin (in bp).}

\item{improve_epsilon}{minimum improve in the objective function to continue the optimization}

\item{min_nuc_prob}{minimum nucleotide probability in every iteration}

\item{unif_prior}{uniform prior for nucleotide probabilities}

\item{is_train}{a boolean vector that determine which subset of sequences to use when screening}

\item{include_response}{include the response in the resulting list (default: TRUE)}

\item{seed}{random seed}

\item{verbose}{show verbose messages.}

\item{kmer_length}{The number of non-gap characters in motifs that will be screened}

\item{...}{
  Arguments passed on to \code{\link[=screen_kmers]{screen_kmers}}
  \describe{
    \item{\code{min_cor}}{Only patterns for which the maximum correlation to one of the response variable is larger than min_cor will be reported}
    \item{\code{min_n}}{Only patterns for which the average number of occurrences in the sequences is larger than min_n will be reported}
    \item{\code{min_gap,max_gap}}{the length of a gap to be considered in the pattern. Only one gap, of length min_gap:max_gap, is being used, and is located anywhere in the motif. Note that this greatly expand the search space (and increase multiple tesing severly).}
    \item{\code{from_range}}{Sequences will be considered only from position from_range (default 0)}
    \item{\code{to_range}}{Sequences will be considered only up to position to_range (default NULL - using the length of the sequences)}
    \item{\code{return_mat}}{Return a matrix of patterns and their correlation to the response variables instead of a data frame. (default: FALSE)}
  }}
}
\value{
a list with the following elements:
\itemize{
\item{pssm: }{data frame with the pssm matrix with the inferred motif, where rows are positions and columns are nucleotides.}
\item{spat: }{a PSSM matrix with the inferred spatial model, where rows are positions and columns are nucleotides.}
\item{pred: }{a vector with the predicted pwm for each sequence.}
\item{response: }{The response matrix. If \code{include_response} is FALSE, the response matrix is not included in the list.}
\item{r2: }{\eqn{r^2} of the prediction with respect to the each response variable.}
\item{ks: }{If \code{score_metric = "ks"}, Kolmogorov-Smirnov test results of the predictions where the response was 1 vs the predictions where the response was 0.}
\item{seed_motif: }{The seed motif that started the regression.}
\item(kmers: ){The k-mers that were screened in order to find the best seed motif (if motif was NULL).}
}
}
\description{
Perform a PWM regression
}
\examples{
res <- regress_pwm(sequences_example, response_mat_example)
res$pssm
res$spat
head(res$pred)

plot_pssm_logo(res$pssm)

}
