% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regress_pwm}
\alias{regress_pwm}
\title{Perform a PWM regression}
\usage{
regress_pwm(
  sequences,
  response,
  motif = NULL,
  bidirect = TRUE,
  epsilon = 0.001,
  spat_min = 0,
  spat_max = NULL,
  min_nuc_prob = 0.001,
  spat_bin = 50,
  min_rms_for_star = 0.001,
  is_train = NULL,
  seed = 60427,
  verbose = FALSE,
  kmer_length = 8,
  ...
)
}
\arguments{
\item{sequences}{A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through \code{toupper})}

\item{response}{A matrix of response variables - number of rows should equal the number of sequences}

\item{motif}{Initial motif to start the regression from. If NULL - a K-mer screen would be performed in order
to find the best kmer for initialization.}

\item{is_train}{a boolean vector that determine which subset of sequences to use when screening}

\item{seed}{random seed}

\item{verbose}{show verbose messages.}

\item{kmer_length}{The number of non-gap characters in motifs that will be screened}

\item{...}{
  Arguments passed on to \code{\link[=screen_kmers]{screen_kmers}}
  \describe{
    \item{\code{min_cor}}{Only patterns for which the maximum correlation to one of the response variable is larger than min_cor will be reported}
    \item{\code{min_n}}{Only patterns for which the average number of occurrences in the sequences is larger than min_n will be reported}
    \item{\code{min_gap,max_gap}}{the length of a gap to be considered in the pattern. Only one gap, of length min_gap:max_gap, is being used, and is located anywhere in the motif. Note that this greatly expand the search space (and increase multiple tesing severly).}
    \item{\code{from_range}}{Sequences will be considered only from position from_range (default 0)}
    \item{\code{to_range}}{Sequences will be considered only up to position to_range (default NULL - using the length of the sequences)}
    \item{\code{return_mat}}{Return a matrix of patterns and their correlation to the response variables instead of a data frame. (default: FALSE)}
  }}
}
\description{
Perform a PWM regression
}
