% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-sample.R
\name{regress_pwm.sample}
\alias{regress_pwm.sample}
\title{Run PWM regression on a sample of the data}
\usage{
regress_pwm.sample(
  sequences,
  response,
  motif_length = 15,
  score_metric = "r2",
  bidirect = TRUE,
  spat_min = 0,
  spat_max = NULL,
  spat_bin = 50,
  improve_epsilon = 0.0001,
  min_nuc_prob = 0.001,
  unif_prior = 0.05,
  is_train = NULL,
  include_response = TRUE,
  seed = 60427,
  verbose = FALSE,
  kmer_length = 6:8,
  min_gap = 0,
  max_gap = 1,
  min_kmer_cor = 0.1,
  consensus_single_thresh = 0.6,
  consensus_double_thresh = 0.85,
  sample_frac = NULL,
  sample_idxs = NULL,
  sample_ratio = 1,
  final_metric = "r2",
  parallel = getOption("prego.parallel", FALSE),
  match_with_db = FALSE,
  motif_dataset = all_motif_datasets(),
  ...
)
}
\arguments{
\item{sequences}{A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through \code{toupper})}

\item{response}{A matrix of response variables - number of rows should equal the number of sequences}

\item{motif_length}{Length of the seed motif. If the motif is shorter than this, it will be extended by wildcards (stars). Note that If the motif is longer than this, it will \emph{not} be truncated.}

\item{score_metric}{metric to use for optimizing the PWM. One of "r2" or "ks". When using "ks" the response variable should be a single vector of 0 and 1.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{spat_min}{start of the spatial model from the beginning of the sequence (in bp)}

\item{spat_max}{end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model
would end at the end of the sequence.}

\item{spat_bin}{size of the spatial bin (in bp).}

\item{improve_epsilon}{minimum improve in the objective function to continue the optimization}

\item{min_nuc_prob}{minimum nucleotide probability in every iteration}

\item{unif_prior}{uniform prior for nucleotide probabilities}

\item{is_train}{a boolean vector that determine which subset of sequences to use when screening}

\item{include_response}{include the response in the resulting list (default: TRUE)}

\item{seed}{random seed}

\item{verbose}{verbosity of the optimization.}

\item{kmer_length}{a vector of kmer lengths to screen in order to find the best seed motif.}

\item{consensus_single_thresh, consensus_double_thresh}{thresholds for the consensus sequence calculation
(single and double nucleotides)}

\item{sample_frac}{fraction of the dataset to sample. When \code{response} is categorical (0 and 1), the sampling would be stratified by the category, i.e. \code{sample_frac} can be a vector of length 2 with the fraction of 0 and 1 responses to sample respectively.
If NULL - the default would be 0.1 for continuous variables, and for binary variables - the number of 0 responses would be equal to \code{sample_ratio} times the number of 1 responses.}

\item{sample_idxs}{indices of the sequences to use. If NULL, the indices would be sampled using \code{sample_frac}.}

\item{sample_ratio}{ratio between the '1' category and the '0' category in the sampled dataset. Relevant only when \code{sample_frac} is NULL.}

\item{final_metric}{metric to use in order to choose the best motif. One of 'ks' or 'r2'. Note that unlike \code{score_metric} which is used in the regression itself, this metric is used only for choosing the best motif out of all the runs on the sampled dataset.}

\item{parallel}{whether to run optimization in parallel. use \code{set_parallel}
to set the number of cores to use.}

\item{match_with_db}{match the resulting PWMs with motif databases using \code{pssm_match}. Note that the closest match
is returned, even if it is not similar enough in absolute terms.}

\item{motif_dataset}{a data frame with PSSMs ('A', 'C', 'G' and 'T' columns), with an additional column 'motif' containing the motif name, for example \code{HOMER_motifs}, \code{JASPAR_motifs} or all_motif_datasets(). By default all_motif_datasets() would be used.}

\item{...}{
  Arguments passed on to \code{\link[=screen_kmers]{screen_kmers}}
  \describe{
    \item{\code{min_cor}}{Only patterns for which the maximum correlation to one of the response variable is larger than min_cor will be reported}
    \item{\code{min_n}}{Only patterns for which the average number of occurrences in the sequences is larger than min_n will be reported}
    \item{\code{min_gap,max_gap}}{the length of a gap to be considered in the pattern. Only one gap, of length min_gap:max_gap, is being used, and is located anywhere in the motif. Note that this greatly expand the search space (and increase multiple tesing severly).}
    \item{\code{from_range}}{Sequences will be considered only from position from_range (default 0)}
    \item{\code{to_range}}{Sequences will be considered only up to position to_range (default NULL - using the length of the sequences)}
    \item{\code{return_mat}}{Return a matrix of patterns and their correlation to the response variables instead of a data frame. (default: FALSE)}
  }}
}
\value{
a list with the following elements:
\itemize{
\item{pssm: }{data frame with the pssm matrix with the inferred motif, where rows are positions and columns are nucleotides.}
\item{spat: }{a data frame with the inferred spatial model, with the spatial factor for each bin.}
\item{pred: }{a vector with the predicted pwm for each sequence.}
\item{consensus: }{Consensus sequence based on the PSSM.}
\item{response: }{The response matrix. If \code{include_response} is FALSE, the response matrix is not included in the list.}
\item{r2: }{\eqn{r^2} of the prediction with respect to the each response variable.}
\item{ks: }{If response is binary, Kolmogorov-Smirnov test results of the predictions where the response was 1 vs the predictions where the response was 0.}
\item{seed_motif: }{The seed motif that started the regression.}
\item{kmers: }{The k-mers that were screened in order to find the best seed motif (if motif was NULL).}
\item{sample_idxs: }{The indices of the sequences that were used for the regression (only for \code{regress_pwm.sample}).}
}
}
\description{
The optimization would be performed with a sampled dataset of size \code{sample_frac} (or explicit sampled indices \code{sample_idxs}) where different candidates of kmers would be regressed in order to find the best seed according to \code{final_metric}.
}
\examples{
res <- regress_pwm.sample(cluster_sequences_example, cluster_mat_example[, 1], final_metric = "ks")
res$pssm
res$spat
head(res$pred)

plot_regression_qc(res)

}
