% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-2phase.R
\name{regress_pwm_two_phase}
\alias{regress_pwm_two_phase}
\title{Run a 2-phase PWM regression}
\usage{
regress_pwm_two_phase(
  sequences,
  response,
  motif_length = 15,
  score_metric = "r2",
  bidirect = TRUE,
  spat_min = 0,
  spat_max = NULL,
  spat_bin = 50,
  improve_epsilon = 0.0001,
  min_nuc_prob = 0.001,
  unif_prior = 0.05,
  is_train = NULL,
  include_response = TRUE,
  seed = 60427,
  verbose = FALSE,
  kmer_length = 5:8,
  min_gap = 0,
  max_gap = 1,
  min_kmer_cor = 0.1,
  two_phase_sample_frac = 0.1,
  first_phase_idxs = NULL,
  first_phase_metric = "r2",
  ...
)
}
\arguments{
\item{sequences}{A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through \code{toupper})}

\item{response}{A matrix of response variables - number of rows should equal the number of sequences}

\item{motif_length}{Length of the seed motif. If the motif is shorter than this, it will be extended by wildcards (stars). Note that If the motif is longer than this, it will \emph{not} be truncated.}

\item{score_metric}{metric to use for optimizing the PWM. One of "r2" or "ks". For categorical response variables (0 and 1), "ks" is recommended, while for continuous response variables, "r2" is recommended. Default is "r2". When using "ks" the response variable should be a single vector of 0 and 1.}

\item{bidirect}{is the motif bi-directional. If TRUE, the reverse-complement of the motif will be used as well.}

\item{spat_min}{start of the spatial model from the beginning of the sequence (in bp)}

\item{spat_max}{end of the spatial model from the beginning of the sequence (in bp). If NULL - the spatial model
would end at the end of the sequence.}

\item{spat_bin}{size of the spatial bin (in bp).}

\item{improve_epsilon}{minimum improve in the objective function to continue the optimization}

\item{min_nuc_prob}{minimum nucleotide probability in every iteration}

\item{unif_prior}{uniform prior for nucleotide probabilities}

\item{is_train}{a boolean vector that determine which subset of sequences to use when screening}

\item{include_response}{include the response in the resulting list (default: TRUE)}

\item{seed}{random seed}

\item{verbose}{show verbose messages.}

\item{kmer_length}{a vector of kmer lengths to screen in order to find the best seed motif.}

\item{two_phase_sample_frac}{fraction of the dataset to sample for the first phase of the optimization (default: 0.1). When \code{response} is categorical (0 and 1), the sampling would be stratified by the category,
i.e. \code{two_phase_sample_frac} can be a vector of length 2 with the fraction of 0 and 1 responses to sample
respectively.}

\item{first_phase_idxs}{indices of the sequences to use in the first phase of the optimization. If NULL, the indices would be sampled using \code{two_phase_sample_frac}.}

\item{first_phase_metric}{metric to use in order to choose the best motif in the first phase of the optimization. One of 'ks' or 'r2'. Note that unlike \code{score_metric} which is used in the regression itself, this metric is used only for choosing the best motif in the first phase of the optimization out of all the runs on the sampled dataset.}
}
\description{
The first phase of the optimization would be performed with a sampled dataset of size \code{two_phase_sample_frac} where different candidates of kmers would be regressed in order to find the best seed. Thenm the optimization would be performed on the full dataset while initializing the motif from the sampled dataset. You can also give explicit indices of the sequences to use in the first phase using \code{first_phase_idxs}.
}
\examples{
res <- regress_pwm_two_phase(sequences_example, response_mat_example, two_phase_sample_frac = 0.1)
res$pssm
res$spat
head(res$pred)

}
