% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{screen_kmers}
\alias{screen_kmers}
\title{Screen for kmers}
\usage{
screen_kmers(
  sequences,
  response,
  kmer_length = 6,
  min_cor = 0.08,
  min_n = 50,
  is_train = NULL,
  min_gap = 0,
  max_gap = 0,
  from_range = 0,
  to_range = NULL,
  return_mat = FALSE,
  seed = 60427,
  verbose = FALSE
)
}
\arguments{
\item{sequences}{A vector of DNA sequences ('A', 'T', 'C' or 'G'. Will go through \code{toupper})}

\item{response}{A matrix of response variables - number of rows should equal the number of sequences}

\item{kmer_length}{The number of non-gap characters in motifs that will be screened}

\item{min_cor}{Only patterns for which the maximum correlation to one of the response variable is larger than min_cor will be reported}

\item{min_n}{Only patterns for which the average number of occurrences in the sequences is larger than min_n will be reported}

\item{is_train}{a boolean vector that determine which subset of sequences to use when screening}

\item{min_gap, max_gap}{the length of a gap to be considered in the pattern. Only one gap, of length min_gap:max_gap, is being used, and is located anywhere in the motif. Note that this greatly expand the search space (and increase multiple testing severely).}

\item{from_range}{Sequences will be considered only from position from_range (default 0)}

\item{to_range}{Sequences will be considered only up to position to_range (default NULL - using the length of the sequences)}

\item{return_mat}{Return a matrix of patterns and their correlation to the response variables instead of a data frame. (default: FALSE)}

\item{seed}{random seed}

\item{verbose}{show verbose messages}
}
\value{
A data frame with the following columns, together with a column for each response variable with the
correlation of the kmers to the response variable:
\itemize{
\item{kmer: }{the kmer pattern, where "*" indicates a wildcard},
\item{max_r2: }{the maximum R^2 to one of the response variables},
\item{avg_n: }{the average number of times the kmer appears in the sequences},
\item{avg_var: }{the variance of the number of times the kmer appears in the sequences},
}
if \code{return_mat} is TRUE, a matrix with correlations to the response variables (where
rows are the kmers) is returned instead of a data frame.
If no kmer is found, an empty data frame is returned.
}
\description{
Screen for kmers
}
\examples{
kmers <- screen_kmers(sequences_example, response_mat_example)
head(kmers)

kmers <- screen_kmers(sequences_example, response_mat_example, return_mat = TRUE)
head(kmers)

kmers <- screen_kmers(sequences_example, response_mat_example, max_gap = 3)
head(kmers)

}
